Expose `Block#fromString`, `LevelParser#value`
